+++ examples/2-replace/23-lisp-lambda.el
@@ -1,0 +1,2 @@
+(lambda (x) (>= x 0))
+(lambda (y) (sqrt y))